BUG: fix windows/32-bit builds

+++ pandas/tseries/tests/test_timeseries.py
@@ -1014,13 +1014,13 @@ class TestTimeSeries(unittest.TestCase):
 
     def test_constructor_int64_nocopy(self):
         # #1624
-        arr = np.arange(1000)
+        arr = np.arange(1000, dtype=np.int64)
         index = DatetimeIndex(arr)
 
         arr[50:100] = -1
         self.assert_((index.asi8[50:100] == -1).all())
 
-        arr = np.arange(1000)
+        arr = np.arange(1000, dtype=np.int64)
         index = DatetimeIndex(arr, copy=True)
 
         arr[50:100] = -1
